moving node native api to new module
