+++ pyrsia_node/src/docker/v2/handlers.rs
@@ -3,7 +3,6 @@
 pub mod blobs;
 
 pub mod manifests;
-pub mod swarm;
 
 // functions available from other modules
